提交初始版本的门窗磁探测器的源码

+++ app/framework/plugin/reporting/reporting.h
@@ -7,7 +7,7 @@
 
 typedef struct {
   uint32_t lastReportTimeMs;
-  uint32_t lastReportValue;
+  uint64_t lastReportValue;
   bool reportableChange;
 } EmAfPluginReportVolatileData;
 extern EmAfPluginReportVolatileData emAfPluginReportVolatileData[];
